fix : 3단계 3차 피드백 반영
