+++ src/main/java/RacingCar/RacingCarApplication.java
@@ -1,5 +1,7 @@
 package RacingCar;
 
+import java.util.Random;
+
 public class RacingCarApplication {
     public static void main(String[] args) {
         InputView inputView = new InputView();
@@ -9,10 +11,11 @@ public class RacingCarApplication {
 
         Racing racing = new Racing(carCount, round);
         ResultView resultView = new ResultView();
+        RandomNumber randomNumber = new RandomNumber();
 
         resultView.resultView();
         for (int i = 0; i < racing.getRound(); i++) {
-            racing.round();
+            racing.round(randomNumber.randomList(carCount));
             resultView.roundView(racing.getCarList());
         }
     }
