+++ src/test/java/RacingCar/CarTest.java
@@ -1,0 +1,4 @@
+package RacingCar;
+
+public class CarTest {
+}
